Add: Now we can use access_right to control user's access

+++ resources/views/admin/user/show.blade.php
@@ -5,10 +5,12 @@
   <h2 class="brand-header">研修生の詳細情報</h2>
 
   <div class="btn-wrapper">
-    {!! Form::open(['route'=>['admin.user.destroy',$user->id],'method'=>'DELETE']) !!}
-      <a class="btn" href="{{ $user->id }}/edit">編集</a>
-      <button class="btn-danger btn" type="submit">削除</button>      
-    {!! Form::close() !!}
+    @if ($selfinfo->access_right & env('ACCESS_RIGHT_USER'))
+      {!! Form::open(['route'=>['admin.user.destroy',$user->id],'method'=>'DELETE']) !!}
+        <a class="btn" href="{{ $user->id }}/edit">編集</a>
+        <button class="btn-danger btn" type="submit">削除</button>
+      {!! Form::close() !!}
+    @endif
   </div>
 
   <div class="container">
@@ -17,22 +19,22 @@
         <h3>ユーザー名</h3>
         {{ $user->name }}
       </li>
-      
+
       <li>
         <h3>苗字</h3>
         {{ $user->info->last_name }}
-      </li>   
+      </li>
 
       <li>
         <h3>名前</h3>
         {{ $user->info->first_name}}
       </li>
-    
+
       <li>
         <h3>性別</h3>
         {{$user->info->sex}}
       </li>
-        
+
       <li>
         <h3>誕生日</h3>
         {{date("Y/m/d", strtotime($user->info->birthday))}}
@@ -42,7 +44,7 @@
         <h3>メールアドレス</h3>
         {{$user->info->email}}
       </li>
-  
+
       <li>
         <h3>電話番号</h3>
         {{$user->info->tel}}
@@ -52,7 +54,7 @@
         <h3>入社日</h3>
         {{date("Y/m/d", strtotime($user->info->hire_date))}}
       </li>
-  
+
       <li>
         <h3>店舗名</h3>
         {{$user->info->store->name}}
